Normalize text to fix french
