+++ cw_bitcoin/pubspec.yaml
@@ -28,7 +28,6 @@ dependencies:
       url: https://github.com/cake-tech/bitbox-flutter.git
       ref: Add-Support-For-OP-Return-data
   rxdart: ^0.27.5
-  unorm_dart: ^0.2.0
   cryptography: ^2.0.5
   bitcoin_base:
     git:
