+++ cw_core/pubspec.yaml
@@ -20,6 +20,7 @@ dependencies:
   intl: ^0.18.0
   encrypt: ^5.0.1
   socks5_proxy: ^1.0.4
+  unorm_dart: ^0.2.0
 #  tor:
 #    git:
 #      url: https://github.com/cake-tech/tor.git
