+++ lib/monero/cw_monero.dart
@@ -221,7 +221,7 @@ class CWMonero extends Monero {
           required int height,
           required String mnemonic}) =>
       MoneroRestoreWalletFromSeedCredentials(
-          name: name, password: password, height: height, mnemonic: mnemonic);
+          name: name, password: password, height: height, mnemonic: normalizeText(mnemonic));
 
   @override
   WalletCredentials createMoneroNewWalletCredentials({
